pyproject.toml
[tool.poetry]
name = "ravintolasovellus"
version = "0.1.0"
description = ""
authors = ["sanikk <sanikk@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
invoke = "^2.2.0"
python-dotenv = "^1.0.1"
flask = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
folium = "^0.16.0"
jupyter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
